deployments/fdp-to-consumable-product/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-pipeline-ref-cdp"
version = "1.0.13"
description = "GCP Pipeline Reference: Generic CDP Transformation — dbt models for FDP to CDP (JOIN pattern across all 3 FDP tables)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["gcp", "bigquery", "dbt", "cdp", "transformation", "reference", "example"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "dbt-bigquery>=1.5.0",
    "gcp-pipeline-core>=1.0.13",
    "gcp-pipeline-transform>=1.0.13",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/enrichmeai/gcp-pipeline-reference"
Documentation = "https://github.com/enrichmeai/gcp-pipeline-reference/tree/main/deployments/fdp-to-consumable-product"
Repository = "https://github.com/enrichmeai/gcp-pipeline-reference"
Framework = "https://pypi.org/project/gcp-pipeline-framework/"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dbt*"]

[tool.setuptools.package-data]
"*" = ["*.sql", "*.yml", "*.yaml", "*.csv", "*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
